Move `TempRegistry` into its own module
We are generalising the `TempRegistry` and adding new features in later commits,
such that it deserves its own module space.
